Update to zig 0.13.0
In the process I changed to using a submodule instead of a dependency -
it seems that zig now requires a `build.zig.zon` for any deps thus
updating to tracy 0.11.1 required the submodule.

Got rid of a workaround for the `SourceLocation` line not being
comptime - seems this has been fixed since it was written.

Bumped the version in the process. Not that it mean anything to zig for
the moment.